Fixes #33
